FIX alter delete method in ClientController

+++ src/main/java/com/api/saofranciscoinn/controllers/ClientController.java
@@ -1,7 +1,9 @@
 package com.api.saofranciscoinn.controllers;
 
 import com.api.saofranciscoinn.dtos.ClientDto;
+import com.api.saofranciscoinn.models.booking.BookingModel;
 import com.api.saofranciscoinn.models.client.ClientModel;
+import com.api.saofranciscoinn.services.BookingService;
 import com.api.saofranciscoinn.services.ClientService;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
@@ -19,10 +21,13 @@ import java.util.UUID;
 @RequestMapping("client")
 public class ClientController {
     final ClientService clientService;
+    final BookingService bookingService;
 
-    public ClientController(ClientService clientService){
-        this.clientService=clientService;
+    public ClientController(ClientService clientService, BookingService bookingService) {
+        this.clientService = clientService;
+        this.bookingService = bookingService;
     }
+
     @PostMapping(value = "/save")
     public ResponseEntity<Object> saveClient(@Valid @RequestBody ClientDto clientDto){
         var clientModel= new ClientModel();
@@ -61,8 +66,13 @@ public class ClientController {
 
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Cliente não encontrado!");
         }
-        //verificar se cliente possui reservas
-        else
+        List<BookingModel> bookings = bookingService.findAll()
+                .stream()
+                .filter(booking -> booking.getClient().getId().equals(clientModelOptional.get().getId()))
+                .toList();
+        if(!bookings.isEmpty()){
+            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Cliente possui reservas!");
+        }
             clientService.delete(clientModelOptional.get());
         return ResponseEntity.status(HttpStatus.OK).body("Cliente deletado com sucesso!");
